[refactor] storage.js 추상화

+++ src/components/Form.jsx
@@ -3,14 +3,17 @@ import styled from "styled-components";
 import { useState } from "react";
 import { useDispatch } from "react-redux";
 import { createTodos, deleteTodos } from "../redux/modules/todos";
+import { storage } from "../utils/storage";
 import nextId from "react-id-generator";
+import Button from "./Button";
+import Input from "./Input";
 
 const Form = () => {
   const id = nextId();
   const [text, setText] = useState("");
   const [deadLine, setDeadLine] = useState("");
   const [checkedItems, setCheckedItems] = useState([]);
-  const [query, setQuery] = useState(localStorage.getItem("search"));
+  const [query, setQuery] = useState(() => localStorage.getItem("search"));
   const dispatch = useDispatch();
 
   // 오늘 날짜
@@ -28,10 +31,9 @@ const Form = () => {
     let todoArr = [];
 
     // 로컬스토리지에 추가
-    todoArr = JSON.parse(localStorage.getItem("allTodos")) || [];
+    todoArr = storage.parseToArray("allTodos") || [];
     todoArr.push(req);
-
-    localStorage.setItem("allTodos", JSON.stringify(todoArr));
+    storage.save("allTodos", todoArr);
 
     // API Post
     dispatch(createTodos(req));
@@ -40,33 +42,27 @@ const Form = () => {
   };
 
   const onDeleteHandler = () => {
-    if (checkedItems.length > 0) {
-      if (window.confirm("삭제할까요?") === true) {
-        // 로컬스토리지의 투두들을 리스트로 변환
-        const todosFromLocalStorage = localStorage.getItem("allTodos");
-        const localTodos = JSON.parse(todosFromLocalStorage);
-
-        // 로컬스토리지에서 삭제
-        for (let i = 0; i < checkedItems.length; i++) {
-          const index =
-            localTodos &&
-            localTodos.findIndex((todo) => todo.id === checkedItems[i]);
-          if (index > -1) {
-            localTodos && localTodos.splice(index, 1);
-          }
-
-          // 삭제된 배열을 다시 로컬스토리지에 넣어줌
-          let allTodos = JSON.stringify(localTodos);
-          localStorage.setItem("allTodos", allTodos);
+    if (checkedItems.length > 0 && window.confirm("삭제할까요?") === true) {
+      // 로컬스토리지의 투두들을 리스트로 변환
+      const localTodos = storage.parseToArray("allTodos");
+      // 로컬스토리지에서 삭제
+      for (let i = 0; i < checkedItems.length; i++) {
+        const index =
+          localTodos &&
+          localTodos.findIndex((todo) => todo.id === checkedItems[i]);
+        if (index > -1) {
+          localTodos && localTodos.splice(index, 1);
         }
-
-        // API Delete
-        dispatch(deleteTodos(checkedItems));
-        setCheckedItems([]);
-      } else {
-        setCheckedItems([]);
-        return false;
+        // 삭제된 배열을 다시 로컬스토리지에 넣어줌
+        storage.save("allTodos", localTodos);
       }
+
+      // API Delete
+      dispatch(deleteTodos(checkedItems));
+      setCheckedItems([]);
+    } else {
+      setCheckedItems([]);
+      return false;
     }
   };
 
@@ -79,26 +75,26 @@ const Form = () => {
     <>
       <StForm onSubmit={onSubmitHandler}>
         <StElementsDiv>
-          <StInput
-            type="text"
-            value={text}
+          <Input
+            type={"text"}
             required
-            placeholder="투두를 입력하세요"
+            value={text}
+            placeholder={"투두를 입력하세요"}
             onChange={(e) => setText(e.target.value)}
           />
-          <StInput
+          <Input
             style={{ cursor: "pointer" }}
-            type="date"
+            type={"date"}
             value={deadLine}
             required
             min={todaysDate}
             onChange={(e) => setDeadLine(e.target.value)}
           />
-          <StButton type="submit">등록</StButton>
-          <StButton type="button" onClick={onDeleteHandler}>
+          <Button type={"submit"}>등록</Button>
+          <Button type={"button"} onClick={onDeleteHandler}>
             삭제
-          </StButton>{" "}
-          <StSearchInput
+          </Button>
+          <Input
             type="search"
             value={query || ""}
             onChange={handleSearch}
@@ -106,7 +102,6 @@ const Form = () => {
           />
         </StElementsDiv>
       </StForm>
-
       <TodoList
         checkedItems={checkedItems}
         setCheckedItems={setCheckedItems}
@@ -128,27 +123,4 @@ const StElementsDiv = styled.div`
   justify-content: center;
   margin-left: 10px;
   gap: 20px;
-`;
-const StInput = styled.input`
-  margin: 35px 0 20px 0;
-  width: 200px;
-  height: 30px;
-  border: none;
-  border-radius: 5px;
-`;
-const StButton = styled.button`
-  margin-top: 35px;
-  width: 50px;
-  height: 30px;
-  border: none;
-  cursor: pointer;
-  border-radius: 5px;
-  background: aliceblue;
-`;
-const StSearchInput = styled.input`
-  margin-top: 35px;
-  width: 200px;
-  height: 30px;
-  border: none;
-  border-radius: 5px;
 `;
